RF+ENH: @with_kill_srv fixture to kill_srv in the tests
